Renamed API folder. Fixed urls.py and api/urls.py to pass initial tests. Added database integration lacking actual database and user
