+++ api/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class OurApplicationConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'GroupFive'
+    name = 'api'
