+++ api/views.py
@@ -2,15 +2,14 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated #for user id
-from GroupFive.models import AnalysisTask
-from GroupFive.serializers import AnalysisRequestSerializer
+from api.models import AnalysisTask
+from api.serializers import AnalysisRequestSerializer
 from .tasks import run_analysis_async
 
-
-#analysis task endpoint
 class AnalysisView(APIView):
     permission_classes = [IsAuthenticated]
 
+    #analysis task endpoint
     def post(self, request):
         serializer = AnalysisRequestSerializer(data=request.data)
         serializer.is_valid(raise_exception=True) #deserialize, check correct input and format, raises 400 Bad Request on fail
@@ -29,10 +28,9 @@ class AnalysisView(APIView):
             "status": task.status
         })
 
-#status endpoint
 class StatusView(APIView):
     permission_classes = [IsAuthenticated]
-
+    #status endpoint
     def get(self, request, task_id):
         task = AnalysisTask.objects.get(id=task_id, user=request.user) #user
 
@@ -40,3 +38,4 @@ class StatusView(APIView):
             "status": task.status,
             "summary": task.results if task.status == "COMPLETED" else None
         })
+
